Fix update til at bruge patch.

+++ StorageSystem/API/Controllers/ProductController.cs
@@ -72,8 +72,8 @@ namespace API.Controllers
             }
         }
 
-        [HttpPut("{id}")]
-        public IActionResult Update(int id, [FromBody] Product product)
+        [HttpPut]
+        public IActionResult UpdateOrCreate([FromBody] Product product)
         {
             if (!ModelState.IsValid)
             {
@@ -82,10 +82,37 @@ namespace API.Controllers
 
             try
             {
-                var existingProduct = ProductService.Get(id);
+                if (!ProductService.Update(product))
+                    return Create(product);
+                else
+                    return NoContent();
+            }
+            catch (Exception e)
+            {
+                // If the product is not found, return 404 Not Found
+                if (e.Message.Contains("Sequence contains no elements"))
+                    return NotFound();
 
-                ProductService.Update(existingProduct, product.Price, product.Name, product.Type);
-                return NoContent();
+                Console.WriteLine($"Error in Update: {e.Message}");
+                return StatusCode(500, e.Message);
+            }
+        }
+
+
+        [HttpPatch]
+        public IActionResult Update([FromBody] Product product)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState); // Return 400 Bad Request with validation errors
+            }
+
+            try
+            {
+                if (!ProductService.Update(product))
+                    return NotFound();
+                else
+                    return NoContent();
             }
             catch (Exception e)
             {
@@ -103,7 +130,10 @@ namespace API.Controllers
         {
             try
             {
-                var product = ProductService.Get(id);
+                Product? product = ProductService.Get(id);
+                if (product == null)
+                    return NotFound();
+
                 ProductService.Remove(product);
                 return NoContent();
             }
